requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
ajsonrpc==1.2.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
bottle==0.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.115.14
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.60.0
fsspec==2025.9.0
gast==0.6.0
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
grpcio==1.75.0
grpcio-status==1.62.3
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.11.3
kiwisolver==1.4.9
langchain-core==0.3.76
langchain-google-genai==2.1.12
langsmith==0.4.30
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
narwhals==2.11.0
numpy==1.26.4
ollama==0.6.1
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
pillow==11.3.0
platformio==6.1.18
plotly==6.4.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyelftools==0.32
Pygments==2.19.2
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
ruff==0.13.1
safehttpx==0.1.6
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==3.1.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
wsproto==1.2.0
xmltodict==1.0.2
zstandard==0.25.0
